Coupon Checkout Fix

+++ resources/views/client/pages/view_my_cart.blade.php
@@ -609,7 +609,8 @@
                  "x-csrf-token": $("[name=_token]").val()
              },
          }).done(response => {
-            amount = parseInt(response);
+            amount = parseFloat(response);
+            // console.log(amount);
             inp.val(amount);
            $('#clickpaypal').click();
      });
